Fix bug by removing app/_workspace after zip creation
By including eggspress-chart, we introduce an error during build arising from attempting to import a module that does not exist (because we do not install the required dependencies for Chart modules). To resolve this, we remove app/_workspace upon creating the starter workspace zip or upon detecting files indicating that a workspace already exists.