Revert previous commit and wrap compileContent in try/catch
The solution to catch unrecognized tags using regex is too aggressive and catches valid tags (any tag that is not a part of UserComponents is deemed "unrecognized"). This further has the unintended consequence of replacing user-written content containing "<" and ">" symbols. For now, the patchwork solution is to allow a site to build, but render an empty page for pages containing faulty components