Update: Improve Components (#62)
* Update .gitignore to ignore /my_*

* Modify prebuild.js to improve component copy behavior

Copy subdirectories that contain all other files on which the main component file depends

* Set overflow-x-hidden on div wrapping Markdown content

Resolves an issue resulting in horizontal scroll on page when there is an Eggspress Chart inserted in content

* Fix code styling not applying to code in li tags

* Update custom pages frontmatter template

Have template include author, snippet, and description fields

* Create utils function getUserDataRecursively

getUserDataRecursively wraps getFilesRecursively and allows component developers to pull user data easily (and safely) from my_data folder. Accepts both the filename and a subfolder to look in.

* Update types/Blog

* Replace types that are "| null" to "| undefined"

To align with recent updates to types/Blog

* Update prebuild.js to handle components starting with "#" in name

This allows user to disable a component by prefixing filename with "#"

* Update prebuild.js

* Add eggspress-chart, Discloure components to starter workspace

By default, eggspress-chart is disabled due to longer build time required when it is on (user must remove "#" in my_components/#Chart.tsx)

* Add README.md to app/_workspace/my_data

This allows git to add my_data to starter workspace (git does not index empty directories)

* Update vercel.json

* Swap out `fs-extra` and `glob` for `fs` and `path`

This is to resolve a build error in Vercel that causes builds in new Eggspress projects to fail

* Revert change made to vercel.json

We still want prebuild.js to run before npm install, so reverting previous commit that switched the ordering of these commands

* Improve tolerance for missing data in category paginated pages

Addresses an edge case in which an author supplies a category slug for which there is no category file in my_categories. This commit handles cases when categoryData is undefined.

* Improve error handling when data is missing in Eggspress Chart

This modifies starter workspace version of Eggspress Chart (#Chart.tsx)

* Update README.md to mention components

Specifically, list MDX support in Features and add two links to docs, one on Using Components and the other about Eggspress Chart

* Fix bug by removing app/_workspace after zip creation

By including eggspress-chart, we introduce an error during build arising from attempting to import a module that does not exist (because we do not install the required dependencies for Chart modules). To resolve this, we remove app/_workspace upon creating the starter workspace zip or upon detecting files indicating that a workspace already exists.

* Handle case where my_components does not exist

In prebuild, if my_components does not exist or cannot be read, overwrite UserComponents.tsx with dummy components so that UserComponents.tsx is still valid

* Add margin-y to default Disclosure component

* Handle errors related to unrecognized tags

If an author inserts a component that is not recognized, either because it does not exist or because it could not be set up properly, then rather than halt build, compileContent will replace the unrecognized component with empty tags (keeping children intact) and log an error in build console--this is to prevent sites from failing to build at all due to a faulty component

* Revert previous commit and wrap compileContent in try/catch

The solution to catch unrecognized tags using regex is too aggressive and catches valid tags (any tag that is not a part of UserComponents is deemed "unrecognized"). This further has the unintended consequence of replacing user-written content containing "<" and ">" symbols. For now, the patchwork solution is to allow a site to build, but render an empty page for pages containing faulty components

* Revert all changes made in this PR to compileContent

try/catch does not resolve issues with handling unrecognized components

* Update prebuild.js to create dummy components

Dummy components ensure that there is usually a fallback for components that are malformed or do not exist, possibly due to a typo or an improperly installed component- this ensures that pages will usually build even if an unrecognized component is inserted in content

* Install archiver from prebuild.js conditionally

Only install archiver in set up mode during prebuild so that we can generate the starter workspace .zip file (archiver is also not available in fresh deployments on Vercel)

* Update prebuild to improve handling of unrecognized components

Dummy components now pass children through a div

* Revert changes made to compileContent.tsx

Prior changes are no longer necessary now that we have dummy components set up during prebuild

* Update eggspress-chart #Chart.tsx to defer loading

Via chartjs-plugin-deferred plugin- this is to improve page load performance